+++ lib/badgeville/client.rb
@@ -1,8 +1,8 @@
 module Badgeville
   class Client
     TIMEOUT_SECS = 3
-    HOST         = "sandbox.v2.badgeville.com"
-    PROTOCOL     = "http"
+    HOST         = 'sandbox.v2.badgeville.com'
+    PROTOCOL     = 'https'
 
     include Helpers
     attr_accessor :user, :site, :player_id, :site_id, :timeout, :debug